Refactor logic for modularity between routes and middlewares
