+++ lib/router.ts
@@ -1,32 +1,19 @@
+import express from "express"
 import readDirRecursive from "./utils/readDirRecursive"
 import path from "path"
 import mapRoutes from "./router/map"
 
 const dir = path.join(process.cwd(), "dist/routes")
 const files = readDirRecursive(dir)
-const paths: string[] = []
 
-const routes = files.map(file => {
-    let fileDir = path.parse(file).dir
-    let fileName = path.parse(file).name
-    paths.push(path.posix.join(fileDir, fileName))
+const router = express.Router()
 
-    fileDir = fileDir
-        .split(path.sep)
-        .map(file => {
-            if (file.startsWith("[") && file.endsWith("]"))
-                return `:${file.slice(1, file.length - 1)}`
-            return file
-        })
-        .join("/")
+const { paths, routes } = mapRoutes(files)
 
-    if (fileName.startsWith("[...") && fileName.endsWith("]")) fileName = "*"
-    else if (fileName.startsWith("[") && fileName.endsWith("]"))
-        fileName = `:${fileName.slice(1, fileName.length - 1)}`
-
-    return path.posix.join(fileDir, fileName)
+routes.forEach((route, index) => {
+    const cwd = path.relative(__dirname, dir)
+    const handlers = require(path.join(cwd, paths[index]))
+    handlers.default && router.all(`/${route}`, handlers.default)
 })
-
-const router = mapRoutes(dir, paths, routes)
 
 export default router
